display tag for room
